[Refactor] Move tree logic to setting composable
